Refactor count_special
